New config (#595)
* Use Config instead of Mix.Config

* Use new config/runtime.exs to reduce duplication
