+++ .env.dev
@@ -1,0 +1,8 @@
+BASE_URL=http://localhost:8000
+DATABASE_URL=postgres://postgres:postgres@127.0.0.1:5432/plausible_dev
+CLICKHOUSE_DATABASE_URL=http://127.0.0.1:8123/plausible_dev
+SECRET_KEY_BASE=/njrhntbycvastyvtk1zycwfm981vpo/0xrvwjjvemdakc/vsvbrevlwsc6u8rcg
+DISABLE_SUBSCRIPTION=false
+ENVIRONMENT=dev
+MAILER_ADAPTER=Bamboo.LocalAdapter
+LOG_LEVEL=debug
